Upload code and demonstrations for arXiv:1812.11276
The original code release at https://github.com/yz93/Learn-to-Interpret-Atari-Agents
was a fork of https://github.com/Kaixhin/Rainbow at commit c97b3f7, but instead of
forking the base repository and committing on top of it, the authors decided to
re-upload the code entirely. This made it more difficult to investigate the exact
changes they made. This commit aims to fix that: its diff with 7ee89f0, the tip
of the master branch of the paper repository, is zero, but it also preserves history.